Update info menu + Added 2 new shortcuts for the 2 shortcut buttons: Screen Lock and Move Z.
In info menu, the Custom Machine Name will be displayed in the Status Bar.

Move Z is a function that has just appeared in the latest Prusa firmware. This makes it possible to move the Z axis live (but in relative value compared to the starting point) and allows those who have not printed wheels to manually raise the Z axis, to do so without going through a lot of click and use of rotary knob.